microserviceB/requirements.txt
pandas
pydantic
python-dotenv
confluent-kafka